requirements.txt
asn1crypto==0.24.0
blinker==1.4
Click==7.0
cryptography==2.1.4
enum34==1.1.6
Flask==1.1.1
Flask-Cors==3.0.7
Flask-Mail==0.9.1
gunicorn==19.9.0
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.10.3
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
MIDIUtil==1.2.1
Pillow==6.0.0
psycopg2==2.8.4
pycrypto==2.6.1
PyJWT==1.7.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.12.0
Werkzeug==0.16.0
